Add an example for React
Add a minimal React app with webpack build. The app includes a local Express server that
serves an index.html. The build is meant to mimic a production build in that the
js bundles and sourcemaps are uploaded to AWS S3. The webpack config executes a
git rev-parse to get the commit sha to use as the code version to pass to rollbar.
In production this might not be what you need (git ls-remote maybe?).
You will need AWS and Rollbar accounts. To run the example:

  - Set your aws and rollbar options in `examples/react/webpack.config.babel.js`
  - `$ cd examples/react`
  - `$ npm run build`
  - `$ npm start`
  - open [http://localhost:8000](http://localhost:8000/)

  When the example app is loaded in a browser,
  the app will throw an error, which will be sent to Rollbar.
  You should be able to log in to Rollbar and see the error with stacktrace
  with line numbers that map to the original source.