Replace after emit logic for webpack v4.29+
Webpack v4.29.0 introduced output.futureEmitAssets, which breaks the plugin
when enabled. This is because we no longer have access to sourcemap
contents. To work around this we can tap into the emittedAssets hook and
keeo a map of sourceMaps and the content to be uploaded oncee asseets
are emitted. This should work in webpack 4 and 5 and be end users should
not notice any change in behavior.

fixes #61 and closes #64